+++ lib/commands/run-test-scenario.ts
@@ -60,6 +60,12 @@ export const builder: yargs.CommandBuilder = {
     boolean: true,
     default: false,
   },
+  level: {
+    describe:
+      "validation level. oav runner validate request and response with different strict level. 'validate-request' only validate request should return 2xx status code. 'validate-request-response' validate both request and response.",
+    string: true,
+    default: "validate-request-response",
+  },
   armEndpoint: {
     describe: "ARM endpoint",
     string: true,