+++ lib/testScenario/postmanCollectionRunnerClient.ts
@@ -19,6 +19,7 @@ import {
 import { inject, injectable } from "inversify";
 import { JsonLoader, JsonLoaderOption } from "../swagger/jsonLoader";
 import { setDefaultOpts } from "../swagger/loader";
+import { ValidationLevel } from "./reportGenerator";
 import { SwaggerAnalyzer } from "./swaggerAnalyzer";
 import { DataMasker } from "./dataMasker";
 import { FileLoader } from "./../swagger/fileLoader";
@@ -58,6 +59,7 @@ export interface PostmanCollectionRunnerClientOption extends BlobUploaderOption,
   jsonLoader?: JsonLoader;
   swaggerFilePaths?: string[];
   baseUrl: string;
+  validationLevel?: ValidationLevel;
 }
 
 function makeid(length: number): string {
@@ -310,6 +312,10 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
     if (overwriteVariables !== undefined) {
       types.push("OverwriteVariables");
     }
+    // For post request do not output response log.
+    if (item.request.method === "POST") {
+      types = types.filter((it) => it !== "DetailResponseLog");
+    }
     const testEvent = new Event({
       listen: "test",
       script: {
@@ -469,7 +475,7 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
         {
           collection: this.collection,
           environment: this.collectionEnv,
-          reporters: ["cli", "json", "junit"],
+          reporters: ["cli", "json"],
           reporter: { json: { export: reportExportPath } },
         },
         function (err, summary) {
@@ -521,6 +527,7 @@ export class PostmanCollectionRunnerClient implements TestScenarioRunnerClient {
           enableUploadBlob: this.opts.enableBlobUploader,
           runId: this.opts.runId,
           swaggerFilePaths: this.opts.swaggerFilePaths,
+          validationLevel: this.opts.validationLevel,
         };
         const reportAnalyzer = inversifyGetInstance(NewmanReportAnalyzer, opts);
         await reportAnalyzer.analyze();