+++ lib/testScenario/reportGenerator.ts
@@ -90,6 +90,8 @@ export interface ResponseDiffItem {
   detail: string;
 }
 
+export type ValidationLevel = "validate-request" | "validate-request-response";
+
 export interface ReportGeneratorOption
   extends NewmanReportParserOption,
     TestResourceLoaderOption,
@@ -100,6 +102,7 @@ export interface ReportGeneratorOption
   junitReportPath?: string;
   testScenarioName?: string;
   runId?: string;
+  validationLevel?: ValidationLevel;
 }
 
 @injectable()
@@ -129,6 +132,7 @@ export class ReportGenerator {
       blobConnectionString: "",
       testDefFilePath: "",
       runId: uuid.v4(),
+      validationLevel: "validate-request-response",
     });
     const swaggerFileAbsolutePaths = this.opts.swaggerFilePaths!.map((it) => path.resolve(it));
     this.exampleQualityValidator = ExampleQualityValidator.create({
@@ -196,11 +200,15 @@ export class ReportGenerator {
           ])
         ).map((it) => _.omit(it, ["exampleName", "exampleFilePath"]));
         const correlationId = it.response.headers["x-ms-correlation-request-id"];
+        const responseDiffResult: ResponseDiffItem[] =
+          this.opts.validationLevel === "validate-request-response"
+            ? await this.exampleResponseDiff(generatedExample, matchedStep)
+            : [];
         this.swaggerExampleQualityResult.stepResult.push({
           exampleFilePath: generatedExample.exampleFilePath,
           operationId: it.annotation.operationId,
           runtimeError,
-          responseDiffResult: this.exampleResponseDiff(generatedExample, matchedStep),
+          responseDiffResult: responseDiffResult,
           stepValidationResult: roundtripErrors,
           correlationId: correlationId,
           statusCode: it.response.statusCode,
@@ -312,19 +320,20 @@ export class ReportGenerator {
     };
   }
 
-  private exampleResponseDiff(
+  private async exampleResponseDiff(
     example: GeneratedExample,
     matchedStep: TestStep
-  ): ResponseDiffItem[] {
+  ): Promise<ResponseDiffItem[]> {
     let res: ResponseDiffItem[] = [];
     if (matchedStep?.type === "restCall") {
       if (example.example.responses[matchedStep.statusCode] !== undefined) {
         res = res.concat(
-          this.responseDiff(
+          await this.responseDiff(
             example.example.responses[matchedStep.statusCode]?.body || {},
             matchedStep.responseExpected,
             this.rawReport!.variables,
-            `/${matchedStep.statusCode}/body`
+            `/${matchedStep.statusCode}/body`,
+            matchedStep.operation.responses[matchedStep.statusCode].schema
           )
         );
       }
@@ -332,12 +341,13 @@ export class ReportGenerator {
     return res;
   }
 
-  private responseDiff(
+  private async responseDiff(
     resp: any,
     expectedResp: any,
     variables: any,
-    jsonPathPrefix: string
-  ): ResponseDiffItem[] {
+    jsonPathPrefix: string,
+    schema: any
+  ): Promise<ResponseDiffItem[]> {
     const env = new VariableEnv();
     env.setBatch(variables);
     try {
@@ -356,6 +366,15 @@ export class ReportGenerator {
             message: "",
             detail: "",
           };
+          const jsonPath: string = it.remove || it.add || it.replace;
+          const propertySchema = this.swaggerAnalyzer.findSchemaByJsonPointer(
+            jsonPath,
+            schema,
+            resp
+          );
+          if (propertySchema["x-ms-secret"] === true) {
+            return undefined;
+          }
           if (it.remove !== undefined) {
             ret.code = "RESPONSE_MISSING_VALUE";
             ret.jsonPath = jsonPathPrefix + it.remove;
@@ -373,6 +392,10 @@ export class ReportGenerator {
           } else if (it.replace !== undefined) {
             ret.code = "RESPONSE_INCONSISTENT_VALUE";
             ret.jsonPath = jsonPathPrefix + it.replace;
+            // Ignore diff when propertySchema readonly is true. When property is readonly, it's probably a random generated value which updated dynamically per request.
+            if (propertySchema.readOnly === true) {
+              return undefined;
+            }
             ret.severity = "Error";
             ret.message = `The actual response value is different from example. Path: ${
               ret.jsonPath
@@ -435,13 +458,15 @@ export class ReportGenerator {
       detail: this.dataMasker.jsonStringify(it.response.body),
     };
     const responseObj = this.dataMasker.jsonParse(it.response.body);
-    ret.code = `RUNTIME_ERROR_${it.response.statusCode}`;
+    ret.code = `RUNTIME_ERROR`;
     ret.message = `code: ${responseObj?.error?.code}, message: ${responseObj?.error?.message}`;
     return ret;
   }
 
   public async generateReport() {
-    this.testDefFile = await this.testResourceLoader.load(this.opts.testDefFilePath);
+    if (this.opts.testDefFilePath !== undefined) {
+      this.testDefFile = await this.testResourceLoader.load(this.opts.testDefFilePath);
+    }
     await this.initialize();
     await this.generateTestScenarioResult(this.rawReport!);
     await this.generateExampleQualityReport();