+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # Change Log - oav
 
+## 6/22 2021 2.5.6
+
+- optimise validation logical. Ignore readonly and secret property response check
+- DO NOT output log when request method is POST
+- Support level option. Current two levels: validate-request, validate-request-response. validate-request will only check request parameters. validate-request-response check both request and response.
+- Improve test scenario markdown report format
+- Bugfix: packaging issue for handlebar files
+- Add test scenario JUnit xml report support
+
 ## 6/10/2021 2.5.5
 
 - add markdown report support for test scenario runner